+++ packages/google_maps_flutter/CHANGELOG.md
@@ -1,10 +1,20 @@
 ## 0.5.23
 
-* Add dummy map initialization
+* Android: Add liteModeEnabled option.
 
 ## 0.5.22
 
-* Add liteModeEnabled option.
+* Support Android v2 embedding.
+* Bump the min flutter version to `1.12.13+hotfix.5`.
+* Fixes some e2e tests on Android.
+
+## 0.5.21+17
+
+* Fix Swift example in README.md.
+
+## 0.5.21+16
+
+* Fixed typo in LatLng's documentation.
 
 ## 0.5.21+15